Add logic to ensure input is a whole number
Adding logic in the program that will check to ensure that the input used in the program is a number.

For now, we restrict the program to accept whole numbers (later revisions will be able to handle floats).

Because we need a number in order to have the program run to successful completion, we will exit code 1 if the input is not a number.